northd: Add support for Logical Datapath Groups.
northd will create only one logical flow with a single logical
datapath group instead of several lflows that differs only by a
datapath column.  We're avoiding creation of datapath groups with
only one datapath to save space.  Old 'logical_datapath' column
used instead.

Feature is disabled.  Configuration option to enable it will be
added in the next commit.

Note: ovn-controller highly depnds on fact that logical flow that
uses multicast group is always received after or in the same update
with the corresponding multicast group.  That will not be the case
with datapath groups.  For that reason northd will not use datapath
groups for such logical flows.

Co-authored-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>