northd: Combine router arp flows.
This flow already matches on 'arp.tpa == ip_address', so instead of
setting 'arp.spa' to this 'ip_address' we can just swap values of
'arp.tpa' and 'arp.spa'.

This elimintes the variable from the 'action', allowing us to just
use the list of all IPs in the match.

Current OVN generates one ARP flow for each load balancer IP for
each router port.  So, if we have 100 ports and 1000 IP addresses,
northd will generate 10^5 ARP flows.  With this change it will
generate only 100 of them.

In case of ovn-k8s scale test, this change alone reduced the database
size from 500 to 200 MB.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>