ovn-northd: Avoid creation of identical datapath groups.
Currently, in a case where datapath groups doesn't need to be changed,
but a new flow should be added, northd will create a new identical
datapath group for a new flow.  It happens because we're not looking
for suitable group in Southbound database, but only tracking ones
created on current iteration.

With this change, northd will collect existing Sb DB datapath groups
while checking logical flows by linking them to unique datapath
groups generated on this run.  This way we can reduce number of
datapath groups.  It wasn't really a problem, almost all the flows
in typical use case are using the same datapath group with only
a few of them using separate copies.  More importantly, this change
allows to avoid full comparison of datapath groups if we already
checked this group from the database once and linked it to a new group
generated on this run.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>