northd: Use larger hash bucket locks instead of dp group ones.
Datapath group locks are taken very frequently while generating logical
flows.  Every time one new datapath is added to the group, northd takes
a mutex, updates a bitmap and unlocks the mutex.

The chances for contention on a datapath group lock are very low, but
even without any contention these constant lock/unlock operations may
take more than 20% of CPU cycles.

We have another type of locks in northd - hash bucket locks which
protect portions of the lflows hash map.  The same locks can be used to
protect datapath groups of logical flows in that hash map.  As long as
we're holding the lock for a portion of a hash map where this lflow is
located, it's safe to modify the logical flow itself.

Splitting out a pair of lflow_hash_lock/unlock() functions that can
be used to lock a particular lflow hash bucket.  Once the hash is
calculated, we can now lock it, then preform an initial lflow creation
and all the datapath group updates, and then unlock.  This eliminates
most of the lock/unlock operations while still maintaining a very low
chance of contention.

Testing with ovn-heater and 4 threads shows 10-25% performance
improvement depending on a weight of lflow node processing in a
particular test scenario.

Hash bucket locks are conditional, so the thread safety analysis is
disabled to avoid spurious warnings [1].  Also, since the order of
taking two locks depends on a "random" hash value, no two locks can
be nested, otherwise we'll risk facing ABBA deadlocks.  For that
reason, some of the loops are split in two.

CoPP meters do not affect the hash value, so they do not affect the
locking scheme.

Unfortunately, the build_lrouter_nat_flows_for_lb() is too complex to
benefit from the change right away.  It will be re-worked separately
in the next commits.

[1] https://clang.llvm.org/docs/ThreadSafetyAnalysis.html#no-conditionally-held-locks

Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>