controller: Fix hairpin SNAT flow explosion for the same SNAT IPs case.
It's common to have 'hairpin_snat_ip' to be configured exactly the
same for each and every Load Balancer in a setup.  For example,
ovn-kubernetes does that, setting '169.254.169.5 fd69::5' value
unconditionally for every LB:
  https://github.com/ovn-org/ovn-kubernetes/blob/cd78ae1af4657d38bdc41003a8737aa958d62b9d/go-controller/pkg/ovn/controller/services/loadbalancer.go#L314

Current implementation of ovn-controller will create an OpenFlow rule
for every datapath for each VIP in case of 'hairpin_snat_ip', because
we need to handle the case where LBs with the same VIP are applied to
different datapaths and have different SNAT IPs.

In large scale setups with tens of thousands of LBs and hundreds of
nodes, this generates millions of OpenFlow rules, making ovn-controller
choke and fall into unrecoverable disconnection loop with poll
intervals up to 200 seconds in ovn-heater's density-heavy scenario with
just 120 nodes.  It also generates rules with thousands of conjunctions
that do not fit into a single FLOW_MOD, breaking the OpenFlow
management communication.

Fix that by checking that all the 'hairpin_snat_ip' options on all the
Load Balancers are exactly the same and using just one OpenFlow rule
per VIP if that's the case.  This is possible because even if LBs with
the same VIP are applied to different datapaths, we're still going to
SNAT with the exactly same hairpin SNAT IP.

Ideally, we can only check that 'hairpin_snat_ip' is the same for all
LBs with the same VIP, but that is not necessary for now, as the only
common use case for this option today is ovn-kubernetes.  Can be added
in the future by enhancing the logic in ovn-northd, if necessary.

Note:

There seems to be a separate issue in ovn-controller.  If only Load
Balancer data is updated, I-P processing is going through the 'lb_data'
node.  And handler for that node deletes old flows and adds new ones
for each updated LB.  This triggers OpenFlow flow DEL + ADD.  On DEL
of hairpin reply learning flows in table 68, OVS cleans up all the
learned flows in table 69 as well, because of 'delete_learned' flag.

However, when changes are happening in higher level nodes and
en_lflow_output_run() triggers lflow_run(), lflow_run() doesn't delete
current hairpin flows before re-processing them.  That leads to the
flow MOD instead of DEL + ADD.  Since the learning flow is not deleted,
but modified, that doesn't trigger removal of learned rules in OVS.
This is what happening if 'lb_same_hairpin_ips' flips the value.
The test adjusted to change the 'hairpin_snat_ip' twice, to trigger
an 'lb_data' processing.  Otherwise, old learned flow for the VIP
stays in table 69.

This should not be a huge issue, as it's still a VIP address and it
should not make any harm to have that learned flow.  Also, the value
of 'lb_same_hairpin_ips' doesn't ever flip in current ovn-kubernetes.

Reported-at: https://bugzilla.redhat.com/2171423
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>