northd: Update obsolete datapath groups instead of re-creating.
Every time we add a new switch or router it may result in addition or
removal of a few datapaths to datapath groups.  Every time that
happens, northd will remove obsolete groups, create new groups that
differ by a few datapaths, and update all affected Sb LBs and logical
flows with references to these new groups.  This generates huge
database transactions that touch half of the records in the Southbound
DB making Sb DB server generate huge updates for all connected clients.

All in all, this behavior leads to high CPU usage in all OVN components
on, for example, attempt to add or remove the node to/from a scaled
OVN-Kubernetes cluster.  It also prevents implementing datapath group
incremental processing in ovn-controller.

Update existing obsolete groups already attached to the logical flow
or Sb load balancer instead of re-creating to avoid these issues.

In case of Load Blancer groups, in the current code, we scan all the
Sb records beforehand and add all the non-stale groups to 'dp_groups'.
So, if Sb LB has a group and it's not in the map, it is stale and can
be safely modified, as no other LB is going to use this group.

In case of logical flows, the order of operations in the current code
is different.  First, we create all the groups that we will need, then
we link existing Sb groups to them as we scan all the Sb logical flow
records.  It means that we need an extra search in order to say if the
valid Sb DP group is going to be used by any other logical flow.  We
should not modify if it will be.  However, this should not result in
any extra work, because we can shortcut the search for the flow that
will actually need this group later.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>