github: Update versions of action dependencies (Node.js 20).
checkout@v3, cache@v3, setup-python@v4 and setup-go@v3 are using
outdated Node.js 16 which is now deprecated in GHA [1], so these
actions may stop working soon.

Updating to most recent major versions with Node.js 20.  This stops
GHA from throwing warnings in every build.

[1] https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/

While at it also updating upload-artifact and download-artifact to
the latest versions.

New setup-go@v5 attempts to cache dependencies by default.  However,
the default path it uses is go.sum in the root directory.  This
triggers a warning, since the file doesn't exist:

  Restore cache failed: Dependencies file is not found in
  /home/runner/work/ovn-kubernetes/ovn-kubernetes.
  Supported file pattern: go.sum

Specify a path to all .sum files we have in the repository to make
the setup-go happy.  This should in theory make the builds a touch
faster.  This change is in line with recent changes in ovn-kubernetes
itself.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>