mac-cache: Fix expiration of active MAC binding due to skipped update.
ovn-controller doesn't update timestamps until 3/4 of the expiration
threshold have passed, but then it may wait for another 3/4.  And since
3/4 + 3/4 > 1, northd would remove the entry before ovn-controller
wakes up to refresh it next time.

For example:
Let's say the threshold is 20 seconds and all the entries we create
have constant traffic flow, so should never expire.  Let's see the
timeline that starts at time A:

  1. [A]    We create one mac entry, next statistics request: A + 15.

  2. [A+5]  We create second entry, next request doesn't change and
            is still at A + 15.

  3. [A+15] We send an OF stats request and get a reply.

     a. First mac entry is updated, because dump interval is 15 and
        the original timestamp is A, which is now + some delta,
        i.e., A + 15 <= A + 15.

     b. Second entry is not updated, because (A + 5) + 15 > A + 15.

     c. The next request is scheduled for 'now + 15' (thresholds didn't
        change), so the next request is at A + 30.

  4. [A+25] northd removes the second entry since it wasn't updated and
            the threshold passed.

  5. [A+30] We send an OF stats request and get a reply, but the entry
            is already removed.

In order to fix that we must never skip the update if the next dump
period may go over the threshold.  This can be achieved by choosing the
cooldown period small enough, so together with the dump period they do
not go over the threshold and actually leave some room for all the
processing and the updates.

Let's use 1/2 of the threshold as the dump period and 1/4 as a cooldown
period.  This should fix the issue while still not sending too many
transactions to the database and not requesting the OpenFlow statistics
too often.  1/2 + 1/4 = 3/4, so we still have enough room for all the
processing before the aging threshold.

New logs added for easier debugging of such issues.

A new test is added to demonstrate the issue and verify the fix.

Reported-at: https://issues.redhat.com/browse/FDP-1130
Fixes: e1ab41ee32aa ("controller: Update MAC binding timestamp")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>